Update `Wait` and `Rate` loop modes for switch to Vulkan
This involved refactoring the `RefreshSync` run loop into a set of
modular functions for easier re-use between loop modes.

Note that while each loop mode is now supported, the `Wait` and `Rate`
present modes are not yet optimal. If using the `Wait` or `Rate` we
should attempt to use the `Mailbox` present mode with triple-buffering
in order to avoid blocking when acquiring an image. This is different to
the `RefreshSync` mode as the `RefreshSync` mode is *driven* by the rate
at which swapchain images can be retrieved for each window.

+++ src/frame.rs
@@ -124,7 +124,7 @@ pub struct Frame {
     // The `Id` whose surface the swapchain image is associated with.
     window_id: window::Id,
     // The `nth` frame that has been presented to the window since the start of the application.
-    nth: usize,
+    nth: u64,
     // The index associated with the swapchain image.
     swapchain_image_index: usize,
     // The image to which this frame is drawing.
@@ -148,7 +148,7 @@ impl Frame {
     pub(crate) fn new_empty(
         queue: Arc<Queue>,
         window_id: window::Id,
-        nth: usize,
+        nth: u64,
         swapchain_image_index: usize,
         swapchain_image: Arc<SwapchainImage>,
     ) -> Result<Self, vulkano::OomError> {
@@ -203,7 +203,7 @@ impl Frame {
     /// The `nth` frame for the associated window since the application started.
     ///
     /// E.g. the first frame yielded will return `0`, the second will return `1`, and so on.
-    pub fn nth(&self) -> usize {
+    pub fn nth(&self) -> u64 {
         self.nth
     }
